fix time_after32() macro

+++ sys/dev/pci/drm/include/linux/jiffies.h
@@ -26,7 +26,7 @@ extern volatile unsigned long jiffies;
 #define nsecs_to_jiffies64(x)	(((uint64_t)(x)) * hz / 1000000000)
 #define get_jiffies_64()	jiffies
 #define time_after(a,b)		((long)(b) - (long)(a) < 0)
-#define time_after32(a,b)	((uint32_t)(b) - (uint32_t)(a) < 0)
+#define time_after32(a,b)	((int32_t)((uint32_t)(b) - (uint32_t)(a)) < 0)
 #define time_after_eq(a,b)	((long)(b) - (long)(a) <= 0)
 #define time_before(a,b)	((long)(a) - (long)(b) < 0)
